requirements.txt
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.1
django-allauth==65.3.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
drf-yasg==1.21.8
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.6
fastapi-cli==0.0.7
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.2
fqdn==1.5.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_resources==6.4.5
inflection==0.5.1
ipykernel==6.29.5
ipython==8.30.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mysqlclient==2.2.6
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.1
notebook_shim==0.2.4
numpy==2.2.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.0.0
pipenv==2024.4.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2==2.9.10
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2024.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.21.0
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==75.4.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.41.3
swagger-spec-validator==3.0.4
terminado==0.18.1
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
virtualenv==20.27.1
voila==0.5.8
watchfiles==1.0.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
